:truck: rename: 예외 관리 파일 위치 변경
